Another Pulumi stack, for installing Flux
This new project is for installing Flux on the cluster created by
`../bootstrap`. The path of least (but still considerable) resistance,
is to use the Flux provider with `kubernetes.yaml.ConfigGroup`. I
would do this all in one project, but I wanted to show Pulumi YAML,
and Pulumi YAML cannot make use of ConfigGroup (because they are in an
overlay, meaning extra SDK code, rather than built-in.)

At least this way, I can show what a "normal" Pulumi program looks
like.

The infra program connects to the cluster by getting the `kubeconfig`
output from the other stack.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>