Create a DB instance and service account key
The app needs a database, and a database needs a database instance (in
CloudSQL). These latter take ages to make, so do it during bootstrap.

We will need a Google Cloud service account with adequate permissions
for the operator to use, so create one of those too, and put the key
in an output so it can be made available in a secret.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>